Split some inner classes out of the parser and refactored some class names.
